requirements.txt
PyMuPDF==1.23.7
Pillow==10.4.0
Flet==0.25